livekit-rag-agent/requirements.txt
# Core LiveKit Dependencies
livekit-agents
livekit-api
livekit-protocol

# LiveKit Plugins
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-turn-detector

# RAG and AI
llama-index
llama-index-embeddings-openai
llama-index-llms-openai
openai

# Document Processing
pypdf
python-docx
markdown

# Essential Dependencies
python-dotenv
pydantic-settings
numpy

# ML Dependencies (fixes PyTorch warning)
torch
transformers

# Optional but recommended
pandas
scipy